refactor(Theme): move salon files into respective component dirs
- Move salon/details_panel into DetailsPanel/salon
- Move salon/preset_list into PresetList/salon
- Keep salon/index.ts at Theme level
