+++ frontend/core/unit/DashboardThread/Appearance/Theme/DetailsPanel/TextureBalls.tsx
@@ -5,8 +5,8 @@ import useThemeKV from '~/hooks/useThemeKV'
 import CloseSVG from '~/icons/CloseLight'
 
 import { PRESET_FIELD } from '../constant'
-import useSalon, { cn, cnMerge } from '../salon/details_panel/texture_balls'
 import type { TThemePresetOverwrite, TThemePresetTokens } from '../spec'
+import useSalon, { cn, cnMerge } from './salon/texture_balls'
 
 type TProps = {
   selectedTokens: TThemePresetTokens
