+++ frontend/core/unit/DashboardThread/Appearance/Theme/DetailsPanel/salon/colors/color_item.ts
