+++ frontend/core/unit/DashboardThread/Appearance/Theme/DetailsPanel/salon/colors/index.ts
