+++ frontend/core/unit/DashboardThread/Appearance/Theme/DetailsPanel/salon/index.ts
