+++ frontend/core/unit/DashboardThread/Appearance/Theme/DetailsPanel/salon/page_glow.ts
