+++ frontend/core/unit/DashboardThread/Appearance/Theme/DetailsPanel/salon/reset_to_preset_menu.ts
