+++ frontend/core/unit/DashboardThread/Appearance/Theme/DetailsPanel/salon/setting_row.ts
