+++ frontend/core/unit/DashboardThread/Appearance/Theme/DetailsPanel/salon/texture_balls.ts
