+++ frontend/core/unit/DashboardThread/Appearance/Theme/PresetList/PresetCard.tsx
@@ -5,8 +5,8 @@ import Checker from '~/widgets/Checker'
 
 import { PRESET_FIELD } from '../constant'
 import MiniBars from '../MiniBars'
-import useSalon from '../salon/preset_list/preset_card'
 import type { TThemePresetCardMode, TThemePresetOption } from '../spec'
+import useSalon, { cn } from './salon/preset_card'
 
 type TProps = {
   preset: TThemePresetOption
