+++ frontend/core/unit/DashboardThread/Appearance/Theme/PresetList/salon/rotate.ts
